Export SpinnerLine as calypso-ui component
