+++ packages/calypso-ui/src/index.js
@@ -1,4 +1,5 @@
 export { default as AsyncLoad } from './async-load';
 export { default as ProgressBar } from './progress-bar';
 export { default as ScreenReaderText } from './screen-reader-text';
+export { default as SpinnerLine } from './spinner-line';
 export { default as WebPreview } from './web-preview';
